One home for the spec numbers, and lint and type gates that reach the whole tree
`tandem/thresholds.py` holds every sec 11 and sec 7.3 figure and the shape of a
judgement against one. A threshold has two values, always: what the specification
asks for, which does not vary by host, and what a particular machine is judged
against once an operator relaxes it in `[gates]`. Both travel with every result —
`pass` against the effective target, `meets_spec` against the spec one, and
`relaxed_criteria` naming each row that is green only because of the config. There is
deliberately no way to express "relaxed" that does not also publish the value it
departed from, because the failure this exists to prevent is a green gate being read
as a met specification six weeks later by someone who was not there when the floor
was lowered. It is a leaf that imports nothing from the package; `config`,
`eval.latency` and `backends.mlx_tier1` read their defaults from it so the spec
numbers live in exactly one place, and a default that sat in both the config
dataclass and the gate function cannot drift invisibly.

Ruff is floored at 0.16.2 and capped below 0.17, with `required-version` making a
mismatch an error rather than a green run that proves nothing. 0.16 moved the default
rule set from 59 rules to 413 and `[tool.ruff.lint]` extends that default instead of
naming four hundred rules, so an older binary lints a twentieth of what CI does and
reports success. The select list is only the delta on top, and every family left out
carries its hit count and its reason, because each looks like an obvious win until it
runs here: ANN and D are 1063 and 801 hits that are mypy's job and 800 edits of
docstring style; S is 994, dominated by `assert` in tests, with the members worth
having already in the default set; ERA reads this repository's explanatory comments as
commented-out code; PTH is 24 mechanical rewrites, several through the KV-cache I/O
that is pinned to plain read/write, finding no defect and so buying pure risk. mypy
runs strict over `src/` and deliberately sets no `python_version`, so it checks
against the interpreter it runs under, and CI matrixes both ends of
`requires-python` for lint, types and tests alike.

The rest of the diff is what those two gates found across the tree, together with the
`[gates]` plumbing that now reads the new module. Two hooks keep it that way: one
formats and lints every file a session touches, the other refuses to create a branch.

+++ src/tandem/gateway/context_scale.py
@@ -41,7 +41,7 @@ class ContextScaler:
         """
         if not self.enabled:
             return tokens
-        return int(math.ceil(tokens * self.factor))
+        return math.ceil(tokens * self.factor)
 
     def headroom_tokens(self) -> int:
         """Real tokens left before the harness ought to have compacted."""
